redirect index.tsx to login page for clarity + use env var conf instead of hardcoded url for tests
